pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chronos-archiver"
version = "1.0.0"
authors = [
    {name = "Douglas Araujo", email = "douglas@example.com"},
]
description = "An archival system to download and preserve websites from the Internet Archive's Wayback Machine"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: System :: Archiving",
]
keywords = ["wayback-machine", "archiving", "web-scraping", "internet-archive", "async"]
dependencies = [
    "aiohttp>=3.8.0",
    "asyncio-throttle>=1.0.0",
    "beautifulsoup4>=4.11.0",
    "click>=8.0.0",
    "lxml>=4.9.0",
    "pydantic>=2.0.0",
    "PyYAML>=6.0",
    "redis>=4.5.0",
    "requests>=2.28.0",
    "SQLAlchemy>=2.0.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "flake8>=6.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
]
postgres = [
    "psycopg2-binary>=2.9.0",
]
all = [
    "chronos-archiver[dev,postgres]",
]

[project.urls]
Homepage = "https://github.com/dodopok/ChronosArchiver"
Documentation = "https://github.com/dodopok/ChronosArchiver/blob/main/README.md"
Repository = "https://github.com/dodopok/ChronosArchiver"
Issues = "https://github.com/dodopok/ChronosArchiver/issues"

[project.scripts]
chronos = "chronos_archiver.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["chronos_archiver"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]